Add README, update project setting for release build option.

+++ Dithering/src/util.h
@@ -7,5 +7,5 @@ uint8_t saturated_add(uint8_t val1, int8_t val2)
 	int16_t val2_int = val2;
 	int16_t tmp = val1_int + val2_int;
 
-	return std::min(std::max((int16_t)0, tmp), (int16_t)255);
+	return static_cast<uint8_t>(std::min(std::max((int16_t)0, tmp), (int16_t)255));
 }
